Update README.md
Emphasise that the reproducibility in the CI scenario comes from having a controlled environment (i.e. the only variable is the fish package being tested).

This is also doable without, but most CI setup will install fish at the major version, meaning the minor and patch version are variables and can cause issues.